requirements.txt
pandas 
numpy 
matplotlib
seaborn
scipy
numpyro 
optax
pyro-ppl
jax 
torch 
scikit-learn
pyarrow
transformers
umap-learn
statsmodels
fuzzywuzzy
rapidfuzz
python-Levenshtein